requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
requests>=2.25.0

# Deep learning
torch>=1.9.0

# Optional but recommended
# Uncomment if you want to use these models:
xgboost>=1.5.0
lightgbm>=3.3.0

# Optional for SHAP analysis:
shap>=0.40.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

